# backend/openapi.yml
...
openapi: 3.1.0
info:
  title: FocusMe Cloud API
  version: 1.0.0
  description: |
    Cloud backend API for FocusMe — provides user authentication,
    cross-device plan synchronization, and family dashboard features.

    ## Authentication

    Most endpoints require a Bearer JWT access token in the
    `Authorization` header. Obtain tokens via `/auth/login` or
    `/auth/register`. Access tokens expire after 15 minutes;
    use `/auth/refresh` to rotate.

    ## Rate Limiting

    - Auth routes: **10 requests/minute** per IP
    - All other routes: **100 requests/minute** per IP
    - Exceeded → `429 Too Many Requests` with `Retry-After` header
  contact:
    name: FocusMe Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.focusme.app/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Development

# ── Security Schemes ──────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token obtained from /auth/login or /auth/register.
        Include as `Authorization: Bearer <token>`.

  # ── Reusable Schemas ──────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code constant for programmatic handling
          example: "AUTH_INVALID_CREDENTIALS"
          enum:
            - AUTH_INVALID_CREDENTIALS
            - AUTH_EMAIL_EXISTS
            - AUTH_TOKEN_EXPIRED
            - AUTH_TOKEN_INVALID
            - AUTH_REFRESH_REUSE
            - SYNC_VERSION_CONFLICT
            - SYNC_PLAN_NOT_FOUND
            - FAMILY_INVITE_EXPIRED
            - FAMILY_ALREADY_MEMBER
            - VALIDATION_FAILED
            - RATE_LIMIT_EXCEEDED
            - FORBIDDEN
            - NOT_FOUND
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error message (do not match on this — use code instead)
          example: "Authentication failed: invalid credentials"
        details:
          type: object
          description: Optional structured error context (e.g., field-level validation errors)

    # ── Health ──────────────────────────────────────────────
...
    Plan:
      type: object
      required: [id, user_id, local_id, plan_json, version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        local_id:
          type: string
        plan_json:
          type: object
        version:
          type: integer
        deleted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Sync ────────────────────────────────────────────────
    PushSyncRequest:
      type: object
      required: [device_id, events]
      properties:
        device_id:
          type: string
          description: The sending device's identifier
          example: "device_xyz789"
        events:
          type: array
          items:
            $ref: "#/components/schemas/SyncEventInput"
    SyncEventInput:
      type: object
      required: [event_type, payload]
      properties:
        event_type:
          type: string
          description: Event classification
          example: "plan_created"
        payload:
          type: object
          description: Event-specific data
    PushSyncResponse:
      type: object
      required: [processed, server_time]
      properties:
        processed:
          type: integer
          description: Number of events successfully processed
          example: 5
        server_time:
          type: string
          format: date-time
    PullSyncResponse:
      type: object
      required: [plans, events, server_time]
      properties:
        plans:
          type: array
          items:
            $ref: "#/components/schemas/Plan"
        events:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              event_type:
                type: string
              payload:
                type: object
              created_at:
                type: string
                format: date-time
        server_time:
          type: string
          format: date-time

    # ── Family ──────────────────────────────────────────────
...
paths:
...
  /sync/push:
    post:
      operationId: pushSync
      summary: Batch push local changes to the server
      description: |
        Sends a batch of local change events to the server. Events are
        stored for other devices to pull. Returns the number of events
        processed and the current server time.
      tags: [Sync]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSyncRequest"
      responses:
        "200":
          description: Events processed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSyncResponse"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /sync/pull:
    get:
      operationId: pullSync
      summary: Pull changes since last sync
      description: |
        Returns all plans and sync events modified/created after the
        `since` timestamp. If `since` is omitted, returns all data
        (full sync). Includes `server_time` for the client to use
        as the next `since` value.
      tags: [Sync]
      security:
        - BearerAuth: []
      parameters:
        - name: since
          in: query
          required: false
          description: ISO 8601 timestamp — only return changes after this time
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Sync data retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullSyncResponse"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer

  # ── Family ───────────────────────────────────────────────────
...